#  Port List functions from JavaScript to Elm