This builds off of #150, so that should be merged first.

In the same spirit as #130, I ported several List API functions from JavaScript to pure Elm.  Some notes:

* I kept `scanl` in a separate commit because it is a tiny performance regression (1.55 -> 1.65 seconds on an old laptop for a 100k element list).  The regression is due to the requisite extra `reverse` call.
* I did not port `sortBy` in terms of `sortWith` because that resulted in a larger performance regression (2.5 -> 3.0 seconds on an old laptop for a 100k element list).  The regression is (likely) due to the extra string comparison required by the JavaScript `sortWith` implementation.  If we decide the performance penalty is worth it, I can port it also.
* There are several native functions (`any`, `drop`, `take`, `repeat`) which were not converted, but could be by introducing an additional 1 or 2 native functions to drive looping.  Either a function like `foldlWhile` that handles early termination, and / or a function like `foldn` (or `Range.foldl`) that handles counting without creating intermediary lists (i.e. without `[1..n]`).  I can implement these in a separate experimental branch if anyone just wants to see what they would look like.