@evancz 2015-02-01
This looks great, nice work!

I think the last point about `any`, `drop`, etc. is interesting. It'd be good to move more code into Elm, so I'd be curious to see what this looks like and if it has performance implications.