@jvoigtlaender 2015-02-01
Concerning the "implement some functions using native loops" aspect, there are already some candidate implementations for such loops in https://github.com/TheSeamau5/elm-loop.