@jonathanhefner 2015-02-01
I created three branches to show what looping could be like for `any`, `repeat`, `take`, and `drop`:

* [Using foldn](https://github.com/jonathanhefner/elm-core/tree/experiment-foldn)
  * `any` cannot be implement via `foldn`, and thus is not ported
* [Using foldlWhile](https://github.com/jonathanhefner/elm-core/tree/experiment-foldlWhile)
  * `repeat` cannot be implemented via `foldlWhile`, but a safe, less-efficient, pure-Elm version is included for consideration (note the ** in the benchmark data)
* [Using loopUntil](https://github.com/jonathanhefner/elm-core/tree/experiment-loopUntil)
  * all 4 functions are possible, however the iteration logic was quite confusing to write (I say this as someone who grew up writing C, C#, and Java)


When benchmarking the branches, I first ran the combined tests for just the 4 functions on a 100k element list (that's the first column).  That didn't provide a clear enough picture, so then I isolated the tests for each function (that's the remaining columns).  For `take` and `drop`, there were multiple tests doing similar work, so I pared it down to just one test taking or dropping 99,999 elements from a 100k element list.

branch     | All 4  | `any` | `repeat`  | `take` n-1  | `drop` n-1
-----------|--------|-------|-----------|-------------|-----------
master     | 3.15s  | 1.0s  | 1.4s      | 1.4s        | 1.1s
foldn      | 4.0s   | n/a   | 1.35s     | 1.65s       | 1.15s
foldlWhile | 4.15s  | 1.05s | 1.45s**   | 1.6s        | 1.3s
loopUntil  | 4.65s  | 1.4s  | 1.55s     | 1.65s       | 1.35s